break out org admin and reg user tests into separate files to avoid overloading test names
