+++ test-http/src/test/cve.py
@@ -1,43 +1,19 @@
 import copy
 import datetime as dt
 import json
-import pytest
 import requests
 import time
 import uuid
 from src import env, utils
-from src.utils import (assert_contains, ok_response_contains,
-                       response_contains, response_contains_json)
+from src.utils import ok_response_contains, response_contains_json
 
 CVE_URL = '/api/cve'
-cve_id = 'CVE-1999-0004'
+cve_id = 'CVE-1999-0001'
 cve_id_dne = 'CVE-3000-0001'
 curr_cve_id = 'CVE-2021-0003'
 update_cve_id = create_cve_id = 'CVE-2000-0008'
 reserved_cve_id = 'CVE-2017-5833'
 reject_cve_id = 'CVE-2009-0009'
-
-
-# GET /cve
-def test_get_all_cves(org_admin_headers):
-    """ services api rejects requests for admin orgs """
-    res = requests.get(
-        f'{env.AWG_BASE_URL}{CVE_URL}/',
-        headers=org_admin_headers
-    )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
-
-
-# GET /cve
-def test_get_all_cves(reg_user_headers):
-    """ services api rejects requests for regular users """
-    res = requests.get(
-        f'{env.AWG_BASE_URL}{CVE_URL}/',
-        headers=reg_user_headers
-    )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
 # GET /cve
@@ -74,28 +50,6 @@ def test_get_cve_by_time_modified():
 
 
 # GET /cve/:id
-def test_get_cve(org_admin_headers):
-    """ services api rejects requests for admin orgs """
-    res = requests.get(
-        f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id}',
-        headers=org_admin_headers
-    )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
-
-
-# GET /cve/:id
-def test_get_cve(reg_user_headers):
-    """ services api rejects requests for regular users """
-    res = requests.get(
-        f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id}',
-        headers=reg_user_headers
-    )
-    assert res.status_code == 403
-    response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
-
-
-# GET /cve/:id
 def test_get_cve():
     res = requests.get(
         f'{env.AWG_BASE_URL}{CVE_URL}/{cve_id}',
@@ -112,34 +66,6 @@ def test_get_cve_dne():
     )
     assert res.status_code == 404
     response_contains_json(res, 'error', 'CVE_RECORD_DNE')
-
-
-# POST /cve/:id
-def test_create_cve(org_admin_headers):
-    """ services api rejects requests for admin orgs """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
-        data = json.load(json_file)
-        res = requests.post(
-            f'{env.AWG_BASE_URL}{CVE_URL}/{create_cve_id}',
-            headers=org_admin_headers,
-            json=data
-        )
-        assert res.status_code == 403
-        response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
-
-
-# POST /cve/:id
-def test_create_cve(reg_user_headers):
-    """ services api rejects requests for regular users """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
-        data = json.load(json_file)
-        res = requests.post(
-            f'{env.AWG_BASE_URL}{CVE_URL}/{create_cve_id}',
-            headers=reg_user_headers,
-            json=data
-        )
-        assert res.status_code == 403
-        response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
 # POST /cve/:id
@@ -213,34 +139,6 @@ def test_create_existent_cve():
         )
         assert res.status_code == 400
         response_contains_json(res, 'error', 'CVE_RECORD_EXISTS')
-
-
-# PUT /cve/:id
-def test_update_cve_record(org_admin_headers):
-    """ services api rejects requests for admin orgs """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
-        data = json.load(json_file)
-        res = requests.put(
-            f'{env.AWG_BASE_URL}{CVE_URL}/{update_cve_id}',
-            headers=org_admin_headers,
-            json=data
-        )
-        assert res.status_code == 403
-        response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
-
-
-# PUT /cve/:id
-def test_update_cve_record(reg_user_headers):
-    """ services api rejects requests for admin orgs """
-    with open('./src/test/CVE-2000-0008_public.json') as json_file:
-        data = json.load(json_file)
-        res = requests.put(
-            f'{env.AWG_BASE_URL}{CVE_URL}/{update_cve_id}',
-            headers=reg_user_headers,
-            json=data
-        )
-        assert res.status_code == 403
-        response_contains_json(res, 'error', 'SECRETARIAT_ONLY')
 
 
 # PUT /cve/:id
